refactor: isolate back compat module graph in its own module
