+++ packages/vite/src/node/plugin.ts
@@ -12,7 +12,8 @@ export type { PluginContext } from 'rollup'
 import type { ConfigEnv, ResolvedConfig, UserConfig } from './config'
 import type { ServerHook, ViteDevServer } from './server'
 import type { IndexHtmlTransform } from './plugins/html'
-import type { EnvironmentModuleNode, ModuleNode } from './server/moduleGraph'
+import type { EnvironmentModuleNode } from './server/moduleGraph'
+import type { ModuleNode } from './server/mixedModuleGraph'
 import type { HmrContext, HotUpdateContext } from './server/hmr'
 import type { PreviewServerHook } from './preview'
 import type { DevEnvironment } from './server/environment'
