+++ packages/vite/src/node/server/hmr.ts
@@ -16,7 +16,8 @@ import { isExplicitImportRequired } from '../plugins/importAnalysis'
 import { getEnvFilesForMode } from '../env'
 import { withTrailingSlash, wrapId } from '../../shared/utils'
 import type { Plugin } from '../plugin'
-import type { EnvironmentModuleNode, ModuleNode } from './moduleGraph'
+import type { EnvironmentModuleNode } from './moduleGraph'
+import type { ModuleNode } from './mixedModuleGraph'
 import type { DevEnvironment } from './environment'
 import { restartServerWithUrls } from '.'
